+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Added support for Cosmos DB Service related cmdlets.
 
 ## Version 1.8.2
 * Added support for partition key and id paths to be part of client encryption policy.